+++ formats/vformats-xml/xmlformat-doc.c
@@ -36,7 +36,7 @@ static osync_bool to_xml(char *input, unsigned int inpsize, char **output, unsig
 	return TRUE;
 }
 
-static void destroy(char *input, size_t inpsize)
+static void destroy(char *input, unsigned int inpsize)
 {
 	free(input);
 }
